refactor: Instructions to README refactor Book.js and SearchBooks.js
Basic instructions to run the app were added to the README file.
On Book.js added a validation to put a placeholder image when the API returns no image.
On SearchBooks.js added the usage of Debounce component from 'react-throttle' to ensure
the event handler for the input field change runs only after the last typed letter.